Move session middleware near session
